pyproject.toml
[tool.poetry]
name = "135-modern-logging"
version = "0.1.0"
description = ""
authors = ["eduardo.diez <eduardo.diez@coinscrapfinance.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
envyaml = "^1.10.211231"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
